application-integration/container/ecs/ecs-efs-demo/requirements.txt
boto3
aws_cdk.aws_ecs==1.41.0
aws_cdk.aws_ec2==1.41.0
aws_cdk.aws_elasticloadbalancingv2==1.41.0
aws_cdk.aws_iam==1.41.0
aws_cdk.aws_efs==1.41.0
aws_cdk.core==1.41.0
awscli
